+++ huffman-compression/src/main.rs
@@ -1,47 +1,18 @@
-pub mod code_generation;
-pub mod encode;
-pub mod whitespace;
+pub mod encode_decode;
+mod prefix_code_table;
 
-use std::{
-    collections::HashMap,
-    env,
-    error::Error,
-    fs::{self},
-};
+use std::{env, error::Error};
 
-use itertools::Itertools;
+use encode_decode::encode_and_decode;
 
 const SKIP_CHALLENGE_PATH: usize = 1;
 
 fn main() -> Result<(), Box<dyn Error>> {
     let mut args = env::args().skip(SKIP_CHALLENGE_PATH);
 
-    let file_name = args.next().ok_or("failed to get file_name")?;
-    char_count_mapping(file_name)?;
+    let input_file_name = args.next().ok_or("failed to get input_file_name")?;
+
+    let output_file_name = args.next().ok_or("failed to get output_file_name")?;
+    encode_and_decode(input_file_name, output_file_name)?;
     Ok(())
-}
-
-fn char_count_mapping(file_name: impl AsRef<str>) -> Result<HashMap<char, u32>, Box<dyn Error>> {
-    let content = fs::read_to_string(file_name.as_ref())?;
-
-    let mapping = content
-        .chars()
-        .into_grouping_map_by(|&x| x)
-        .fold(0_u32, |acc, _key, _value| acc + 1);
-
-    Ok(mapping)
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::char_count_mapping;
-
-    #[test]
-    fn test_char_count_mapping() {
-        let file_name = "test-135-0.txt";
-        let mut mapping = char_count_mapping(file_name).expect("Read file provided");
-
-        assert_eq!(mapping.remove(&'X'), Some(333));
-        assert_eq!(mapping.remove(&'t'), Some(223000));
-    }
 }
